Add files via upload
This repository shows day-to-day commands( syntax) used during analysis, it is an expression on how to use JOINS to relate tables, FULL, INNER, RIGHT, LEFT, FUNCTIONS, DELETE, UPDATE, WHERE Filtering, how to IMPORT a dataset into Postgres and database and use of CASE, LIKE cards, and so on, this is backed up with few selected img. to help you understand the code.